Move editorConfig to its own dir
